backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
pydantic-settings
python-multipart
